+++ zfit/core/binneddata.py
@@ -2,6 +2,7 @@
 from .baseobject import BaseObject
 from .dimension import BaseDimensional
 from .interfaces import ZfitBinnedData
+from .. import z
 from ..util.ztyping import NumericalTypeReturn
 
 
@@ -9,14 +10,19 @@ class BinnedData(BaseDimensional, ZfitBinnedData, BaseObject):  # TODO: add dtyp
 
     def __init__(self, obs, counts, w2error, name: str = "BinnedData"):
         super().__init__(name=name, obs=obs)
-        self._count = self._input_check_counts(counts)
+        self._counts = self._input_check_counts(counts)
         self._w2error = w2error
+
+    @classmethod
+    def from_numpy(cls, obs, counts, w2error, name):
+        counts = z.convert_to_tensor(counts)
+        return cls(obs=obs, counts=counts, w2error=w2error, name=name)
 
     def _input_check_counts(self, counts):  # TODO
         return counts
 
     def get_counts(self, bins) -> NumericalTypeReturn:
-        pass
+        return self._counts
 
     def weights_error_squared(self) -> NumericalTypeReturn:
-        pass
+        return self._w2error